Bugs: Fixed bugs in apriori and linked_list
Function generate_frequent_size_one now takes an absolute
min_support_count rather than a ratio.

Revised ll_get_nth to use a int16_t for the index.  Had encountered wrap
around errors from the use of an uint8_t.

+++ c/src/linked_list.c
@@ -284,7 +284,7 @@ void* ll_search(List *list, void* data)
 }
 
 
-void* ll_get_nth(List *list, uint8_t index)
+void* ll_get_nth(List *list, uint16_t index)
 {
     Node *current;
